kafka_consumer_ingestor/requirements.txt
kafka-python==2.0.2
elasticsearch==8.15.0
